refactor to use tailwind
